Refactor Map Card UI Components (#1679)
* head

* move vars inside function to avoid scope issues

* fix

* extract into smaller components for readability
